fix(stdlib): fix error reporting on incorrect arguments
