+++ src/interpreter/Environment.ts
@@ -1,6 +1,7 @@
 import { Identifier } from "../lexer";
 import { BrsType, RoAssociativeArray, Int32, BrsInvalid, RoSGNode, Callable } from "../brsTypes";
 import { ComponentDefinition } from "../componentprocessor";
+import { Location } from "../lexer/Token";
 import { BrsError } from "../Error";
 
 /** The logical region from a particular variable or function that defines where it may be accessed from. */
